requirements.txt
# =============================================================================
# WanderPlan AI - Python Dependencies
# =============================================================================

# --- Web Framework ---
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==13.0

# --- Data Validation ---
pydantic==2.9.0

# --- Messaging (Kafka) ---
aiokafka==0.11.0

# --- Data Stores ---
redis[hiredis]==5.1.0
motor==3.5.0                    # async MongoDB driver
asyncpg==0.29.0                 # async PostgreSQL driver
sqlalchemy[asyncio]==2.0.35

# --- Vector Store ---
pinecone-client==5.0.0

# --- LLM / Agent Framework ---
anthropic==0.34.0
langchain==0.3.0
langchain-anthropic==0.2.0
langchain-community==0.3.0

# --- External APIs ---
httpx==0.27.0                   # async HTTP client for API integrations

# --- Search ---
elasticsearch[async]==8.13.0

# --- Observability ---
prometheus-client==0.21.0
structlog==24.4.0

# --- Utilities ---
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0  # JWT handling
icalendar==5.0.13                 # Calendar export
